requirements.txt
numpy
requests
flask
pillow
torch
torchvision
service-streamer